Use site id instead of site URL in the login token.
Using of site URL was a leftover from the time when Mozilla Persona
was used for the authentication. Persona tokens included site URL and
required verification of the URL to ensure the token is not intendent
for a different site. This introduced complexity that required
wwwhisper to be aware of all valid site domains (trusting
Host header was not a secure option). With own implementation of login
token verification internal site id can be simply used.