Move more access control logic to the proxy
Proxy no longer makes is-authorized requests to wwwhisper backend for
each incoming HTTP request, but retrieves and caches location and user
data from the backend and uses this data to locally make authorization
decisions. As the result, wwwhisper backend is not contacted for each
incoming request, allowing much faster, fully local processing for
majority of requests.